抽离日志配置；定义API的参数结构；删除多余文件
